+++ src/main/java/com/innowise/rudkovskii/service/UserService.java
@@ -1,6 +1,8 @@
 package com.innowise.rudkovskii.service;
 
 import com.innowise.rudkovskii.entity.User;
+import com.innowise.rudkovskii.exception.ResourceNotFoundException;
+import com.innowise.rudkovskii.exception.ValidationException;
 import com.innowise.rudkovskii.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,12 +17,16 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
-    public void createUser(User user){
-        userRepository.save(user);
+    public User createUser(User user) throws ValidationException {
+        if(userRepository.existsByEmail(user.getEmail())){
+            throw new ValidationException("Email already exists");
+        }
+        return userRepository.save(user);
     }
 
-    public User getById(int id) {
-        return userRepository.findById(id).orElse(null);
+    public User getById(int id){
+        return userRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
     }
 
     public List<User> getAll(){
@@ -28,11 +34,15 @@ public class UserService {
     }
 
     public User getByEmail(String email){
-        return userRepository.findByEmail(email);
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found with email: " + email));
     }
 
-    public void updateUser(int id, User user){
-        userRepository.updateUser(id,
+    public User updateUser(int id, User user){
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException("User not found with id: " + id);
+        }
+        return userRepository.updateUser(id,
                 user.getName(),
                 user.getSurname(),
                 user.getBirthDate(),
@@ -40,6 +50,9 @@ public class UserService {
     }
 
     public void deleteUser(int id){
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException("User not found with id: " + id);
+        }
         userRepository.deleteById(id);
     }
 }
